+++ app/src/main/java/angel/panduro/dev/walletregister/presentation/ui/model/DebtInformation.kt
@@ -2,7 +2,7 @@ package angel.panduro.dev.walletregister.presentation.ui.model
 
 data class DebtInformation(
     val id: Long = 0,
-    val idWallet: Int,
+    val idWallet: Long,
     val nameCard: String,
     val typeMoney: String,
     val debt: Float,
